chore: Update plot_bands_and_hog function to include a max_bands parameter
The plot_bands_and_hog function in utils.py has been updated to include a new max_bands parameter. This parameter allows the user to specify the maximum number of bands to plot in the 2-row layout. The default value is set to 12. This change improves the flexibility and customization of the function.