chore: Update runscripts/basicLauncher.sh to include batch_size parameter
This commit updates the runscripts/basicLauncher.sh file to include a new batch_size parameter in the command line arguments for the SenTrainer.py script. The batch_size parameter is set to 8 for all the seed values (89, 16, 123, 402, 233, 701, 108) and band values (8, 2, 3, 4). This change allows for better control over the batch size used in the training process.